Stop running the script if files already exist

+++ install.sh
@@ -1,19 +1,22 @@
 #!/bin/bash
-cp ./.bash_profile_extras ~/
-cp ./.bashrc_extras ~/
+#/usr/bin/bash
+if [ ! -f ~/.bash_profile_extras ]; then
+	cp ./.bash_profile_extras ~/
+	cp ./.bashrc_extras ~/
 
-# Append to `.bash_profile`
-echo "
+	# Append to `.bash_profile`
+	echo "
 
-if [ -f ~/.bash_profile_extras ]; then
-	. ~/.bash_profile_extras
+	if [ -f ~/.bash_profile_extras ]; then
+		. ~/.bash_profile_extras
+	fi
+	if [ -f ~/.bashrc_extras ]; then
+		. ~/.bashrc_extras
+	fi"  | tee -a ~/.bash_profile
+
+	echo "# configuration for customization options per environment
+
+	alias ..code=''" >> ~/.bash_custom
+
+	source ~/.bash_profile
 fi
-if [ -f ~/.bashrc_extras ]; then
-	. ~/.bashrc_extras
-fi"  | tee -a ~/.bash_profile
-
-echo "# configuration for customization options per environment
-
-alias ..code=''" >> ~/.bash_custom
-
-source ~/.bash_profile
